# src/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"ItemDTO": {
				"properties": {
					"uuid": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"category": {
						"type": "string"
					},
					"count": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"uuid",
					"name",
					"category",
					"count"
				],
				"type": "object",
				"additionalProperties": false
			},
			"InventoryCreationParams": {
				"properties": {
					"name": {
						"type": "string"
					},
					"category": {
						"type": "string"
					},
					"count": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"category",
					"count"
				],
				"type": "object",
				"additionalProperties": false
			},
			"InventoryEditParams": {
				"properties": {
					"uuid": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"category": {
						"type": "string"
					},
					"count": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"uuid"
				],
				"type": "object",
				"additionalProperties": false
			},
			"InventoryDeleteParams": {
				"properties": {
					"uuid": {
						"type": "string"
					}
				},
				"required": [
					"uuid"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ShipmentDTO": {
				"properties": {
					"uuid": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"items": {
						"items": {
							"$ref": "#/components/schemas/ItemDTO"
						},
						"type": "array"
					}
				},
				"required": [
					"uuid",
					"name",
					"items"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ShipmentCreationParams": {
				"properties": {
					"name": {
						"type": "string"
					},
					"items": {
						"items": {
							"$ref": "#/components/schemas/ItemDTO"
						},
						"type": "array"
					}
				},
				"required": [
					"name",
					"items"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "inventory-tracking",
		"version": "1.0.0",
		"description": "Web application that implements basic CRUD functionality in addition to undeletion.",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Edward Chang"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/api/inventory": {
			"get": {
				"operationId": "GetItem",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ItemDTO"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "uuid",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateItem",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ItemDTO"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/InventoryCreationParams"
							}
						}
					}
				}
			},
			"put": {
				"operationId": "EditItem",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ItemDTO"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/InventoryEditParams"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteItem",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/InventoryDeleteParams"
							}
						}
					}
				}
			}
		},
		"/api/inventory/items": {
			"get": {
				"operationId": "GetItems",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ItemDTO"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "category",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/shipment": {
			"post": {
				"operationId": "CreateShipment",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ShipmentDTO"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ShipmentCreationParams"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetShipment",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ShipmentDTO"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "uuid",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}